requirements.txt
accelerate==0.25.0
faiss-cpu==1.7.4
nltk==3.8.1
numpy==1.25.2
pypdf2==3.0.1
rank-bm25==0.2.2
sentence-transformers==2.2.2
streamlit==1.27.0
torch==2.0.1
transformers==4.34.0